# Manually inline decodeVarint