… bench.txt
… sstable/bench.txt