Decoding varints occurs multiple times in some of the inner-most code of
sstable seeking and iteration. Manually inlining the varint decoding
provides a significant speedup on both go1.11 and go1.12 (the numbers
below are for go1.11). Perhaps someday this won't be necessary, but
until then the perf win is worth grabbing.

```
name                          old time/op  new time/op  delta
BlockIterSeekGE/restart=16-8   488ns ± 1%   416ns ± 1%  -14.81%  (p=0.000 n=10+10)
BlockIterSeekLT/restart=16-8   531ns ± 1%   460ns ± 1%  -13.25%  (p=0.000 n=10+10)
BlockIterNext/restart=16-8    22.4ns ± 0%  18.1ns ± 1%  -19.42%  (p=0.000 n=6+10)
BlockIterPrev/restart=16-8    38.9ns ± 3%  32.1ns ± 2%  -17.52%  (p=0.000 n=10+9)
TableIterSeekGE/restart=16-8  1.32µs ± 0%  1.20µs ± 1%   -9.72%  (p=0.000 n=8+10)
TableIterSeekLT/restart=16-8  1.37µs ± 1%  1.24µs ± 0%   -9.50%  (p=0.000 n=9+10)
TableIterNext/restart=16-8    25.6ns ± 1%  20.8ns ± 3%  -18.58%  (p=0.000 n=10+10)
TableIterPrev/restart=16-8    40.8ns ± 2%  36.2ns ± 3%  -11.30%  (p=0.000 n=10+10)
```