@petermattis 2019-04-13
Added a few more commits with additional micro optimizations to `blockIter`. These micro optimizations add up. Here is the delta from before these changes (i.e. `master`):

```
name                          old time/op  new time/op  delta
BlockIterSeekGE/restart=16-8   487ns ± 1%   352ns ± 1%  -27.75%  (p=0.000 n=9+10)
BlockIterSeekLT/restart=16-8   538ns ± 2%   406ns ± 1%  -24.63%  (p=0.000 n=10+10)
BlockIterNext/restart=16-8    22.3ns ± 1%  12.7ns ± 0%  -43.35%  (p=0.000 n=10+9)
BlockIterPrev/restart=16-8    37.1ns ± 1%  27.8ns ± 3%  -25.11%  (p=0.000 n=10+10)
TableIterSeekGE/restart=16-8  1.33µs ± 1%  1.09µs ± 1%  -17.85%  (p=0.000 n=9+9)
TableIterSeekLT/restart=16-8  1.38µs ± 1%  1.15µs ± 1%  -16.46%  (p=0.000 n=10+8)
TableIterNext/restart=16-8    25.3ns ± 1%  16.0ns ± 1%  -36.95%  (p=0.000 n=9+10)
TableIterPrev/restart=16-8    42.1ns ± 3%  32.9ns ± 4%  -21.83%  (p=0.000 n=10+10)
```