@petermattis 2019-04-17
> I am surprised functions like DecodeInternalKey are not inlineable.

I'd guess it is just over whatever threshold of size the Go compiler uses as a heuristic.

> Hope our DB doesn't evolve into one giant function someday :).

Heh, I don't think that will happen. The inner loops like this are rare. I wish there was a slightly cleaner way to express that inlining should happen. I was even pondering adding a pre-processing step to accomplish this, but that seemed like overkill.