@jordanlewis 2019-04-20
Why do we use varints at all? Decoding varints is one of the most expensive things in the hot decode paths of CockroachDB. I don't know the context here, but could we consider replacing the use of varints in most places with fixed size integers?