@petermattis 2019-04-20
> Why do we use varints at all?

Because the sstable format that we want to be compatible with uses varints.

> Decoding varints is one of the most expensive things in the hot decode paths of CockroachDB. I don't know the context here, but could we consider replacing the use of varints in most places with fixed size integers?

I've experimented with both fixed size integers (`uint16`) and other varint encodings. Neither show any benefit over the inlining done here. The other varint encoding I experimented with can be significantly faster for multi-byte varints:

```
// xxxxxxx1  7 bits in 1 byte
// xxxxxx10 14 bits in 2 bytes
// xxxxx100 21 bits in 3 bytes
// xxxx1000 28 bits in 4 bytes
// xxx10000 35 bits in 5 bytes
```

The low-bits describe how full bytes follow. Benchmarking shows this is only a win when a varint needs 3 or more bytes to be encoded. For 1 and 2 byte varints the inlining done in this PR is either faster or the same speed.